pyproject.toml
[tool.poetry]
name = "smart-health-cms"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Django = "^4.2"
djangorestframework = "^3.14.0"
python-dotenv = "^1.0.0"


[tool.poetry.group.dev.dependencies]
ruff = "^0.0.284"
black = "^23.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
